// packages/gte-ts/src/internal/generated/types.gen.ts
export type OrderRejectReason = 'invalid_session' | 'invalid_account' | 'invalid_market' | 'invalid_side' | 'invalid_price' | 'invalid_quantity' | 'insufficient_allowance' | 'order_not_found' | 'market_closed' | 'rate_limit_exceeded' | 'invalid_signature' | 'unauthorized_signer' | 'signature_expired' | 'invalid_nonce' | 'no_allowance' | 'allowance_expired' | 'invalid_order' | 'book_at_capacity';

export type OrderStatus = 'new' | 'partially_filled' | 'filled' | 'cancelled' | 'rejected' | 'expired' | 'replaced' | 'pending_new';

export type OrderType = 'market' | 'limit' | 'stop_limit' | 'stop_market' | 'liquidation';

export type PositionSide = 'long' | 'short';

export type Side = 'buy' | 'sell';

export type TimeInForce = 'gtc' | 'ioc' | 'fok' | 'gtd';
// ...
export type CreateOrdersResponse = {
    results: Array<OrderResult>;
};
// ...
export type OrderResult = {
    orderId: string;
    clientOrderId: string;
    status: OrderStatus;
    timestamp: string;
    error?: string | null;
    rejectReason: OrderRejectReason;
};